requirements.txt
# Core data processing
pyspark==3.5.1
delta-spark==3.1.0

# Data warehouse
duckdb==0.10.2

# Data quality
great-expectations==0.18.13

# Testing
chispa==0.9.4
pytest==8.1.1
pytest-cov==5.0.0

# Code quality
ruff==0.4.1
mypy==1.10.0

# Data generation
faker==24.11.0

# Utilities
pyyaml==6.0.1
jsonschema==4.22.0
python-dotenv==1.0.1

# Type stubs
types-PyYAML==6.0.12.20240311
